Correção do tipo decimal para float em DB2TableDefiner

+++ modules/definers/db2.py
@@ -10,7 +10,7 @@ class DB2TableDefiner(BaseTableDefiner):
         'CHARACTER': 'string',
         'CLOB': 'text',
         'DATE': 'date',
-        'DECIMAL': 'decimal',
+        'DECIMAL': 'float',
         'DOUBLE': 'double',
         'INTEGER': 'integer',
         'LONG VARCHAR': 'text',
